# openapi_schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CosyVoice API",
    "description": "Cross-lingual Voice Cloning API with CosyVoice2 (v2) and CosyVoice3 (v3)",
    "version": "3.0.0",
    "contact": {
      "name": "CosyVoice API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8012",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v3/voices/": {
      "get": {
        "tags": ["Voice Management (v3)"],
        "summary": "List voices",
        "operationId": "list_voices_v3",
        "parameters": [
          {
            "name": "voice_type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/VoiceType"
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of voices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Voice Management (v3)"],
        "summary": "Create voice",
        "operationId": "create_voice_v3",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/VoiceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voice created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceResponse"
                }
              }
            }
          },
          "409": {
            "description": "Voice already exists"
          }
        }
      }
    },
    "/api/v3/voices/{voice_id}": {
      "get": {
        "tags": ["Voice Management (v3)"],
        "summary": "Get voice",
        "operationId": "get_voice_v3",
        "parameters": [
          {
            "name": "voice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voice details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Voice not found"
          }
        }
      },
      "put": {
        "tags": ["Voice Management (v3)"],
        "summary": "Update voice",
        "operationId": "update_voice_v3",
        "parameters": [
          {
            "name": "voice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voice updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Voice Management (v3)"],
        "summary": "Delete voice",
        "operationId": "delete_voice_v3",
        "parameters": [
          {
            "name": "voice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voice deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "VoiceType": {
        "type": "string",
        "enum": ["sft", "zero_shot", "cross_lingual", "instruct"]
      },
      "AudioFormat": {
        "type": "string",
        "enum": ["wav", "mp3", "flac"],
        "default": "wav"
      },
...
      "VoiceCreateRequest": {
        "type": "object",
        "required": ["voice_id", "name", "voice_type", "audio_file"],
        "properties": {
          "voice_id": {
            "type": "string",
            "description": "Unique voice identifier"
          },
          "name": {
            "type": "string",
            "description": "Human-readable voice name"
          },
          "description": {
            "type": "string",
            "description": "Voice description"
          },
          "voice_type": {
            "$ref": "#/components/schemas/VoiceType"
          },
          "language": {
            "type": "string",
            "description": "Primary language (e.g., zh, en, ja)"
          },
          "prompt_text": {
            "type": "string",
            "description": "Text matching the audio sample"
          },
          "audio_format": {
            "$ref": "#/components/schemas/AudioFormat"
          },
          "audio_file": {
            "type": "string",
            "format": "binary",
            "description": "Audio file for voice cloning"
          }
        }
      },
      "VoiceUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "language": {
            "type": "string"
          }
        }
      },
      "VoiceResponse": {
        "type": "object",
        "properties": {
          "voice_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "voice_type": {
            "$ref": "#/components/schemas/VoiceType"
          },
          "language": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "audio_format": {
            "$ref": "#/components/schemas/AudioFormat"
          },
          "file_size": {
            "type": "integer"
          },
          "duration": {
            "type": "number"
          },
          "sample_rate": {
            "type": "integer"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
      "VoiceListResponse": {
        "type": "object",
        "properties": {
          "voices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoiceResponse"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}